+++ DeepSkyStacker/DSSProgress.h
@@ -15,17 +15,17 @@ public :
 	};
 	virtual ~CDSSProgress() {};
 
-	virtual void	SetNrUsedProcessors(LONG lNrProcessors=1) {};
+	virtual void	SetNrUsedProcessors(int lNrProcessors=1) {};
 	void	SetJointProgress(bool bJointProgress)
 	{
 		m_bJointProgress = bJointProgress;
 	};
 	virtual void	GetStartText(CString & strText) = 0;
 	virtual void	GetStart2Text(CString & strText) = 0;
-	virtual	void	Start(LPCTSTR szTitle, LONG lTotal1, bool bEnableCancel = true) = 0;
-	virtual void	Progress1(LPCTSTR szText, LONG lAchieved1) = 0;
-	virtual void	Start2(LPCTSTR szText, LONG lTotal2) = 0;
-	virtual void	Progress2(LPCTSTR szText, LONG lAchieved2) = 0;
+	virtual	void	Start(LPCTSTR szTitle, int lTotal1, bool bEnableCancel = true) = 0;
+	virtual void	Progress1(LPCTSTR szText, int lAchieved1) = 0;
+	virtual void	Start2(LPCTSTR szText, int lTotal2) = 0;
+	virtual void	Progress2(LPCTSTR szText, int lAchieved2) = 0;
 	virtual void	End2() = 0;
 	virtual bool	IsCanceled() = 0;
 	virtual bool	Close() = 0;
